# swagger.yaml
openapi: 3.0.0
info:
  title: Chatty Web Architect API
  version: 1.0.0
  description: API documentation for Chatty Web Architect backend.
servers:
  - url: http://localhost:3001/api
paths:
...
  /chat/message:
    post:
      summary: Send message to AI
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                message:
                  type: string
                personality:
                  type: string
                conversationId:
                  type: string
      responses:
        '200':
          description: AI response
...
  /chat/conversations/{id}:
    get:
      summary: Get specific conversation
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Conversation details
    delete:
      summary: Delete conversation
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Conversation deleted
    patch:
      summary: Update conversation
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Conversation updated
...